% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{DataTable}
\alias{DataTable}
\title{DataTable dash-table-experiments component}
\usage{
DataTable(id = NULL, editable = TRUE, filterable = FALSE,
  sortable = TRUE, columns = NULL, row_selectable = FALSE,
  selected_row_indices = list(), enable_drag_and_drop = NULL,
  header_row_height = NULL, min_height = NULL, min_width = NULL,
  row_height = NULL, row_scroll_timeout = NULL, tab_index = NULL,
  filters = list(), rows = NULL, row_update = NULL, sortColumn = NULL,
  sortDirection = NULL, setProps = NULL)
}
\arguments{
\item{id}{}

\item{editable}{}

\item{filterable}{}

\item{sortable}{}

\item{columns}{Order of columns. Note that the column names are specified in `rows` but without order. This attribute allows you to specify a custom order for your columns.}

\item{row_selectable}{}

\item{selected_row_indices}{}

\item{enable_drag_and_drop}{}

\item{header_row_height}{}

\item{min_height}{}

\item{min_width}{}

\item{row_height}{}

\item{row_scroll_timeout}{}

\item{tab_index}{}

\item{filters}{}

\item{rows}{}

\item{row_update}{}

\item{sortColumn}{}

\item{sortDirection}{}

\item{setProps}{}
}
