% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{VirtualizedTable}
\alias{VirtualizedTable}
\title{VirtualizedTable dash-table-experiments component}
\usage{
VirtualizedTable(id = NULL, changed_data = NULL, dataframe = NULL,
  display_index = TRUE, index_name = "", editable = TRUE,
  styles = list(container = list(`font-size` = "1rem"), cell =
  list(`white-space` = "nowrap", `border-left` = "'thin lightgrey solid'",
  `border-bottom` = "'thin lightgrey solid'", `border-top` = "none",
  `border-right` = "none", `text-overflow` = "ellipsis", `overflow-x` =
  "hidden", `box-sizing` = "border-box"), `cell-rightmost` = list(`border-right`
  = "'thin lightgrey solid'"), header = list(`text-align` = "left",
  `border-left` = "'thin lightgrey solid'", `border-bottom` =
  "'thin lightgrey solid'", `border-top` = "'thin lightgrey solid'",     
  `font-weight` = 600, display = "inline-block", `box-sizing` = "border-box"),
  `header-rightmost` = list(`border-right` = "'thin lightgrey solid'")),
  column_width = 80, row_height = 20, height = 500, width = 600,
  setProps = NULL)
}
\arguments{
\item{id}{The ID of the component, used to identify the component in Dash callbacks}

\item{changed_data}{}

\item{dataframe}{The data of this table, in the form of a Pandas DataFrame.to_dict('split') For example, {    'columns': ['Column 1', 'Column 2', 'Column 3'],    'data': [[1, 2, 3], [4, 3, 1]],    'index': [1, 2, 3] }  If `editable=True`, then this `dataframe` will be updated when the user changes values}

\item{display_index}{Whether or not to show the index column}

\item{index_name}{The name of the index column. This name is not included in the `dataframe` object}

\item{editable}{Whether or not this table is editable}

\item{styles}{CSS styles for all of the different elements in this table. The `styles` have a default value. If you want to remove all default styles, set `styles` to `{}`.}

\item{column_width}{Column width (in pixels)}

\item{row_height}{Row height (in pixels)}

\item{height}{Height of entire table (in pixels)}

\item{width}{Width of entire table (in pixels)}

\item{setProps}{Dash supplied function}
}
