% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components.R
\name{DataTable}
\alias{DataTable}
\title{DataTable component}
\usage{
DataTable(id = NULL, editable = TRUE, filterable = FALSE,
  sortable = TRUE, resizable = TRUE, column_widths = NULL,
  columns = NULL, row_selectable = FALSE, selected_row_indices = list(),
  enable_drag_and_drop = NULL, header_row_height = NULL,
  min_height = NULL, min_width = NULL, max_rows_in_viewport = NULL,
  row_height = 35, row_scroll_timeout = NULL, tab_index = NULL,
  filters = list(), rows = NULL, row_update = NULL, sortColumn = NULL,
  sortDirection = NULL, setProps = NULL)
}
\arguments{
\item{id}{}

\item{editable}{Are the cells in the table editable? If `True`, you can listen to `rows` or `row_update` to get the updated data.}

\item{filterable}{Should the filtering UI in the Table appear? If `True`, you can listen to `rows` or `row_update` to get the updated data.}

\item{sortable}{Is the table sortable? If `True`, click on the column headers to sort by that column. If `True`, you can listen to `rows` or `row_update` to get the updated data.}

\item{resizable}{Are the columns resizble? If `True`, then the columns can be resized by clicking and dragging on the border on the edge of the column header. If `False`, they cannot be resized.}

\item{column_widths}{Column widths (in pixels) of each column}

\item{columns}{Order of columns. Note that the column names are specified in `rows` but without order. This attribute allows you to specify a custom order for your columns.}

\item{row_selectable}{}

\item{selected_row_indices}{}

\item{enable_drag_and_drop}{}

\item{header_row_height}{}

\item{min_height}{}

\item{min_width}{}

\item{max_rows_in_viewport}{Maximum number of of rows to display in the viewport. If `max_rows_in_viewport` is 10, then if there are 15 rows, only 10 will be displayed in the viewport but if there are 5 rows, then the viewport will shrink to only show those 5 rows. This takes precedence over `min_height`.}

\item{row_height}{}

\item{row_scroll_timeout}{}

\item{tab_index}{}

\item{filters}{}

\item{rows}{}

\item{row_update}{}

\item{sortColumn}{}

\item{sortDirection}{}

\item{setProps}{}
}
